+++ CHANGELOG.md
@@ -7,6 +7,11 @@
 
 ## 1.47.0
 
+## 1.46.1
+
+- Deprecated `\Wizaplace\SDK\Company\CompanyService::registerC2CCompany` use `CompanyService::register` instead
+- Added `\Wizaplace\SDK\Company\CompanyC2CRegistration`
+
 ## 1.46.0
 
 ### New features